rename: CategoryComponents 파일을 page와 style 분리
하나의 파일에 작성되있던 style부분을 분리하고
폴더구조에 맞게 이동시킴
